gaussian_splatting: extract scene-director sphere-effector cluster into its own TU
Decomposition of the GaussianSplatSceneDirector god-file (the second-largest in
the module). Moves the five sphere-effector coordination member methods
(_build_sorted_sphere_effector_payload, build_sphere_effector_payload_for_renderer,
get_scene_effector_debug_state_for_instance, get_sphere_effector_count_for_renderer,
get_sphere_effector_generation_for_renderer) out of gaussian_splat_scene_director.cpp
into a new partial-class TU core/scene_director_sphere_effectors.cpp (the proven
partial-class pattern; declarations stay in the header, so the live
get_scene_effector_debug_state_for_instance caller in gaussian_splat_node_3d.cpp
is unaffected).

The file-static _get_effector_scope_specificity helper is shared with a method
that stays behind, so it was hoisted (not moved) to a new minimal
core/scene_director_internal.h as `static inline`; both TUs include it. The
const_cast-on-SharedWorld + world_mutex reliance in _build_sorted_* is preserved
verbatim (not "fixed").

Result: gaussian_splat_scene_director.cpp 2976 -> 2645 LOC. Pure code move,
behavior-preserving.

Note: this branch and chore/gs-remove-dead-effector-queries both edit
gaussian_splat_scene_director.cpp off master; apply them in sequence at
integration (trivial conflict).

Risk: R1/R2. Evidence: full editor build compiles + links clean
(scons ... dev_build=yes tests=yes, exit 0); guards green
(run_module_tests.py --guard-only, 108 OK).
